part2: countries: move data fetching to custom hooks
