+++ part2/countries/src/components/Country.jsx
@@ -1,36 +1,9 @@
-import { useEffect, useState } from "react";
+import Weather from "./Weather";
 
-import countriesService from "../services/countries";
-import weatherService from "../services/weather";
+import { useFetchCountryData } from "../hooks/useFetchCountryData";
 
 const Country = ({ countryName }) => {
-  const [country, setCountry] = useState(null);
-  const [weather, setWeather] = useState(null);
-  const [error, setError] = useState(null);
-
-  useEffect(() => {
-    countriesService
-      .get(countryName)
-      .then((data) => setCountry(data))
-      .catch((error) => {
-        setError(
-          `The following error occured while retrieving country data from the server: '${error}'.`
-        );
-      });
-  }, [countryName]);
-
-  useEffect(() => {
-    if (country) {
-      weatherService
-        .get({ lat: country?.latlng[0], lon: country?.latlng[1] })
-        .then((data) => setWeather(data))
-        .catch((error) => {
-          setError(
-            `The following error occured while retrieving weather data from the server: '${error}'.`
-          );
-        });
-    }
-  }, [country]);
+  const [country, weather, error] = useFetchCountryData(countryName);
 
   if (!country) return null;
   if (error) return <>{error}</>;
@@ -50,13 +23,12 @@ const Country = ({ countryName }) => {
         ))}
       </ul>
       <img src={flags.png} alt={flags.alt} />
-      <h2>weather in {capitalName}</h2>
-      <div>temperature {weather.main.temp} Celcius</div>
-      <img
-        src={`https://openweathermap.org/img/wn/${weather.weather[0].icon}.png`}
-        alt={weather.weather[0].main}
-      />
-      <div>wind {weather.wind.speed} m/s</div>
+      {weather && (
+        <>
+          <h2>weather in {capitalName}</h2>
+          <Weather weather={weather} />
+        </>
+      )}
     </>
   );
 };
