+++ part2/countries/src/components/CountryList.jsx
@@ -1,23 +1,43 @@
+import { useState } from "react";
+
 import Country from "./Country";
+import Filter from "./Filter";
 
-const CountryList = ({ countries, setCountries }) => {
-  if (countries?.length >= 10)
-    return <div>Too many matches, specify another filter</div>;
+import { useFetchCountries } from "../hooks/useFetchCountries";
 
-  if (countries?.length === 0)
-    return <div>No countries matching your search criteria</div>;
+const CountryList = () => {
+  const [filter, setFilter] = useState("");
+  const [countries, error] = useFetchCountries(filter);
 
-  if (countries?.length === 1)
-    return <Country countryName={countries[0].name.common} />;
+  const handleFilterChange = (event) => setFilter(event.target.value);
+
+  let component;
+
+  if (countries?.length === 0) {
+    component = <div>No countries matching your search criteria</div>;
+  } else if (countries?.length === 1) {
+    component = <Country countryName={countries[0].name.common} />;
+  } else if (countries?.length >= 10) {
+    component = <div>Too many matches, specify another filter</div>;
+  } else {
+    component = countries?.map((country) => (
+      <div key={country.name.official}>
+        {country.name.official}
+        <button onClick={() => setFilter(country.name.common)}>show</button>
+      </div>
+    ));
+  }
+
+  if (error) return <>{error}</>;
 
   return (
     <>
-      {countries?.map((country) => (
-        <div key={country.name.official}>
-          {country.name.official}
-          <button onClick={() => setCountries([country])}>show</button>
-        </div>
-      ))}
+      <Filter
+        text="find countries"
+        filter={filter}
+        handleFilterChange={handleFilterChange}
+      />
+      {component}
     </>
   );
 };
